Update my scene and prefabs

+++ Team 3 Game Mechanics Project/Assets/Scripts/Personal Folders/Harry H/Health.cs
@@ -6,7 +6,7 @@ public class Health : MonoBehaviour
 {
     private bool IsDead;
     [SerializeField] public int MaxHealth;
-    [HideInInspector] public float currentHealth;
+    [HideInInspector] public int currentHealth;
     private int healthDifference;
     private int value;
     public float damageCooldownTime;
